requirements.txt
apturl==0.5.2
asgiref==3.2.3
asn1crypto==0.24.0
beautifulsoup4==4.8.2
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
confusable-homoglyphs==3.2.0
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==2.0
django-bootstrap4==1.1.1
django-cleanup==4.0.0
django-heroku==0.3.1
django-registration==2.4.1
django-rest-framework==0.1.0
django-tinymce==2.8.0
djangorestframework==3.11.0
gunicorn==20.0.4
httplib2==0.9.2
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pbr==5.4.4
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psycopg2==2.8.4
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyperclip==1.7.0
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.6.1
python-debian==0.1.32
python-decouple==3.3
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
soupsieve==1.9.5
sqlparse==0.3.0
system-service==0.3
systemd-python==234
testresources==2.0.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==16.7.8
wadllib==1.3.2
whitenoise==5.0.1
xkit==0.0.0
zope.interface==4.3.2
